requirements.txt
anthropic>=0.39.0
Pillow>=10.0.0
python-telegram-bot>=20.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
